+++ R/Definitions.R
@@ -17,7 +17,8 @@ initiateRstoxData <- function(){
 	signifDigits <- 6
 	
 	# Define the time format used by Stox formats:
-	StoxDateTimeFormat <- "%Y-%m-%d %H:%M:%OS"
+	#StoxDateTimeFormat <- "%Y-%m-%d %H:%M:%OS"
+	StoxDateTimeFormat <- "%Y-%m-%dT%H:%M:%OS3Z"
 	StoxTimeZone <- "UTC"
 	
 	# Get the path to the extdata folder:
@@ -58,7 +59,21 @@ initiateRstoxData <- function(){
 	)
 	
 	# Define the columns required for VariableConversionTable:
-	StoxBioticTranslationRequiredColumns <- c("VariableName", "Value", "NewValue")
+	TranslationRequiredColumns <- c("VariableName", "Value", "NewValue")
+	
+	# Define the ICESBiotic keys (check with the package author whether this is already defined when reading the data):
+	ICESAcousticKeys <- list(
+		Cruise =   "LocalID", 
+		Log =    c("LocalID", "Distance"), 
+		Sample = c("LocalID", "Distance", "ChannelDepthUpper"), 
+		Data =   c("LocalID", "Distance", "ChannelDepthUpper", "SaCategory")
+	)
+	ICESBioticKeys <- list(
+		Cruise =    "LocalID", 
+		Haul =    c("LocalID", "Gear", "Number"), 
+		Catch =   c("LocalID", "Gear", "Number", "SpeciesCode", "SpeciesCategory"), 
+		Biology = c("LocalID", "Gear", "Number", "SpeciesCode", "SpeciesCategory", "StockCode", "FishID")
+	)
 	
 	#### Assign to RstoxDataEnv and return the definitions: ####
 	definitionsNames <- ls()